requirements.txt
crewai
streamlit
python-dotenv
openai
tavily-python